+++ lua/obsidian/commands/next_week.lua
@@ -1,0 +1,4 @@
+return function()
+  local note = require("obsidian.weekly").next_week()
+  note:open()
+end